requirements.txt
asyncio==3.4.3
python-dotenv==1.1.0
requests==2.32.3
telegram==0.0.1
